Added Maven. Cleaned up Code etc.
