+++ src/main/java/me/szilprog/simplenpc/utils/WaitingMessageType.java
@@ -1,4 +1,4 @@
-package me.szilprog.simplenpc;
+package me.szilprog.simplenpc.utils;
 
 public enum WaitingMessageType {
     NONE, DISPLAY_NAME, SKIN_NAME, COMMAND, COOLDOWN, PERMISSION, PMESSAGE
